Handle View Contexts, Improve General Structure
This commit got out of hand but I'm going to explain it all in here.

Firstly, this tracks the currently selected cluster, in the event that
we're in "Top Down" view this will be null. From this information we can
track the current context. If any node-related (Street View) events
occur which relate to a cluster other than the selected cluster then
they will be ignored.

Additionally this creates some helpers for enabling and disabling
various listeners such that they can be toggled based on current
context. Just some DRY.

Also there was a one line function for slot listening for a single click
that was only called from one place, removed the unnecessary lines and
obfuscation from that.